fix major bugs on the implementation of the final user's flow logic, including purchase order backend and frontend logic and invoice backend and frontend logic. Add mail test server for docuseal. Add docuseal backup for testing use case with pre-defined docuseal contract template

+++ backend/app/schemas/pedido_compra_schema.py
@@ -4,12 +4,12 @@ from datetime import datetime
 from enum import StrEnum
 
 
-class PedidoCompraBase(BaseModel): 
+class PedidoCompraBase(BaseModel):
     status: str = Field(default='Pendente', description = 'status atual do pedido de compra')
     data_assinatura: datetime = Field(..., description = 'timestamp de assinatura do pedido de compra')
     info: str = Field(default='esign utilizando Docuseal', description = 'informacoes adicionais do pedido de compra e esign')
     esign_id: str = Field(default='Pendente', description = 'identificador do contrato entre a empresa e fornecedor')
-    
+
     class Config:
         from_attributes = True
 
@@ -17,7 +17,7 @@ class PedidoCompraGet(PedidoCompraBase):
     pk_id_pedido_compra: int = Field(..., description = 'identificador unico de um pedido de compra')
     titulo_requisicao: str = Field(..., description = 'titulo de uma requisicao')
     razao_social: str = Field(..., description = 'razao social de um fornecedor')
-    preco_total: int = Field(..., description = 'preco total da proposta')
+    preco_total: float = Field(..., description = 'preco total da proposta')
 
 class CreatePedidoCompra(PedidoCompraBase):
     fk_id_gerente: int = Field(..., description = 'identificador unico do gerente associado ao pedido de compra')
